src/backend/synthetic_data/requirements.txt
# Interfaz de usuario
gradio>=4.0.0

# Clientes de API
openai>=1.3.0
anthropic>=0.5.0
huggingface-hub>=0.19.0

# Procesamiento de datos
pandas>=2.0.0
pyarrow>=14.0.0  # Para soporte de Parquet
numpy<2.0.0  # Force numpy 1.x for compatibility

# Model inference
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.24.0
sentencepiece>=0.1.99

# Utilidades
python-dotenv>=1.0.0
rich>=13.0.0
tqdm>=4.65.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Desarrollo
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0 